changed app.tsx structure, added todos, realised autoLogin
