+++ src/entities/Todo/model/todoType.ts
@@ -1,0 +1,9 @@
+export type TodoType = {
+	_id: string
+	title: string
+	order: number
+	completed: boolean
+	description: string
+	createdAt: string
+	updatedAt: string
+}
