+++ src/entities/User/model/userType.ts
@@ -1,0 +1,4 @@
+export type UserType = {
+	access_token: string
+	username: string
+}
